requirements.txt
APScheduler==3.6.3
cryptography==41.0.7
Flask==3.1.1
pandas==2.0.3
pika==1.3.2
pygtrans==1.6.1
pymongo==4.13.0
python-dotenv==1.1.0
python-telegram-bot==13.15
pytz==2023.3
qrcode==7.4.2
Requests==2.32.3
tronpy==0.5.0
